[RFR] Only apply monkeypatching when the module is installed
