+++ base_technical_features/models/__init__.py
@@ -1,2 +1,2 @@
-from . import basemodel
+from . import basemodel_monkeypatch
 from . import ir_ui_menu
